feat: FPX_TRANSPORT supports aiohttp and httpx

+++ fpx/route/ticket.py
@@ -11,11 +11,9 @@ from sanic.server.websockets.impl import WebsocketImplProtocol
 from sqlalchemy.orm.query import Query
 from webargs_sanic.sanicparser import use_kwargs
 
-from fpx import utils
+from fpx import exception, schema, utils
 from fpx.model import Ticket
 from fpx.pipes import Pipe
-
-from .. import schema, exception
 
 log = logging.getLogger(__name__)
 
@@ -33,14 +31,17 @@ async def index(request: request.Request, page: int) -> response.HTTPResponse:
             "page": page,
             "count": base.count(),
             "tickets": [t.for_json() for t in q],
-        }
+        },
     )
 
 
 @ticket.route("/generate", methods=["POST"], ctx_requires_client=True)
 @use_kwargs(schema.TicketGenerate(), location="json")
 async def generate(
-    request: request.Request, type: str, items: Any, options: dict[str, Any]
+    request: request.Request,
+    type: str,
+    items: Any,
+    options: dict[str, Any],
 ) -> response.HTTPResponse:
     ticket = Ticket(type=type, items=items, options=options)
     if request.app.config.FPX_NO_QUEUE:
@@ -58,22 +59,21 @@ async def download(request: request.Request, id: str):
     ticket = db.query(Ticket).get(id)
 
     if ticket is None:
-        raise exception.NotFound({"id": "Ticket not found"})
+        raise exception.NotFoundError({"id": "Ticket not found"})
 
     if not ticket.is_available:
-        raise exception.NotAuthorized(
-            {"access": "You must wait untill download is available"}
+        raise exception.NotAuthorizedError(
+            {"access": "You must wait untill download is available"},
         )
 
     response = await request.respond()
-    assert response
+    if not response:
+        raise ValueError("response")
 
     async with Pipe.choose(ticket, request) as pipe:
         response.content_type = pipe.content_type()
         filename = pipe.filename()
-        response.headers[
-            "content-disposition"
-        ] = f'attachment; filename="{filename}"'
+        response.headers["content-disposition"] = f'attachment; filename="{filename}"'
 
         with utils.ActiveDownload(request.app.ctx.active_downloads, id):
             db.delete(ticket)
@@ -91,21 +91,17 @@ async def wait(request: request.Request, ws: WebsocketImplProtocol, id: str):
     active = request.app.ctx.active_downloads
     ticket = db.query(Ticket).get(id)
     if ticket is None:
-        raise exception.NotFound({"id": "Ticket not found"})
+        raise exception.NotFoundError({"id": "Ticket not found"})
 
     def _position():
         offset = q.index(id) if id in q else len(q)
-        return (
-            offset
-            + len(active)
-            - request.app.config.SIMULTANEOURS_DOWNLOADS_LIMIT
-        )
+        return offset + len(active) - request.app.config.SIMULTANEOURS_DOWNLOADS_LIMIT
 
     async def send_position(position):
         await ws.send(
             json.dumps(
-                {"position": position, "available": ticket.is_available}
-            )
+                {"position": position, "available": ticket.is_available},
+            ),
         )
 
     position = _position()
@@ -114,12 +110,12 @@ async def wait(request: request.Request, ws: WebsocketImplProtocol, id: str):
         db.commit()
     await send_position(position)
     if ticket.is_available:
-        log.debug(f"Ticket {id} is already available. Closing connection")
+        log.debug("Ticket %s is already available. Closing connection", id)
         await ws.close(reason="available")
         return
 
     if id in q:
-        log.debug(f"{id} already in queue as position {q.index(id)}: {q}")
+        log.debug("%s already in queue as position %s: %s", id, q.index(id), q)
         await ws.send(json.dumps({"error": "Already waiting"}))
         log.debug("Closing connection")
         await ws.close(reason="aleready in queue")
@@ -137,11 +133,11 @@ async def wait(request: request.Request, ws: WebsocketImplProtocol, id: str):
             await ws.close(reason="available")
 
     q.append(id)
-    log.debug(f"Appended {id} to download queue: {q}")
+    log.debug("Appended %s to download queue: %s", id, q)
 
     try:
         await ws.wait_for_connection_lost(timeout=3600)
     except (WebsocketClosed, asyncio.CancelledError):
         q.remove(id)
-        log.debug(f"Removed {id} from download queue: {q}")
+        log.debug("Removed %s from download queue: %s", id, q)
         raise
